ImageData FromBytes and FromUint8Array speed up

+++ SpawnDev.BlazorJS/JSObjects/TypedArrays/Uint8Array.cs
@@ -92,10 +92,10 @@ namespace SpawnDev.BlazorJS.JSObjects
         /// <param name="sourceBytes"></param>
         public Uint8Array(byte[] sourceBytes) : base(JS.ReturnMe<IJSInProcessObjectReference>(sourceBytes)) { }
         /// <summary>
-        /// The Uint32Array() constructor creates Uint32Array objects.
+        /// The Uint8Array() constructor creates Uint8Array objects.
         /// </summary>
         /// <param name="array"></param>
-        public Uint8Array(Array<byte> array) : base(JS.New(nameof(Uint32Array), array)) { }
+        public Uint8Array(Array<byte> array) : base(JS.New(nameof(Uint8Array), array)) { }
         #endregion
         /// <summary>
         /// Extracts a section of an array and returns a new array. See also Array.prototype.slice().
